+++ arch/arm/include/cxd56xx/nxaudio.h
